wlan: Link speed should reflect HT40 and SGI
It was reported that the link speed displayed in the Wifi GUI did not
reflect HT40 or SGI setting of the connected AP.  Analysis of the code
showed that SGI was being considered, but HT40 was not.  Update the
code to extract the HT40 setting from the statistics returned from
WCNSS.  Note that there is a companion change to WCNSS to properly
report the HT40 and LGI settings.  In addition update the logic used
to determine and cache the max speed since the max speed can change
over time.

Change-Id: Ie5300294142962b31debb0e121b19cfea60bbaa5
CRs-Fixed:  354157