wlan: Update OBSS config to Riva
OBSS feature enable/disable config is passed to Riva.  This field was
already there in HAL message, however WDI was not populating it.
Depending on this feature, Riva will decide whether to filter all the
beacons in SoftAP mode.  By default Riva is reading this config as
enabled, and because of this host is recieving beacons in SoftAP
mode.

Change-Id: I73535ea26a999cc90c36fb554078c5f8201c572e
CRs-Fixed: 331169