wlan: Fixes to address concurrency scenarios for Bt AMP
To address the following failing scenarios:
1. File transfer does not fall back to BR/EDR and terminates the
   transfer with Wi-Fi functionality
2. Enabling Wi-Fi hotspot on the remote device while AMP transfer is
   in progress, crashes the sender

The root cause for #1 was the deregister with HCI was happening even
before AMP disconnect event is sent to HCI. The fix is the make sure
disconnect is sent before deregister.

The root cause for #2 was we had pkts in WLAN stack on Rx while client
(HCI) has already deregistered, causing a NULL pointer access. The fix
is to add sanity check on rx path to take care of this scenario.

Change-Id: I5d76348e4f79a2a7716a93effebc85449a1b55eb
CRs-Fixed: 355459, 355467