wlan: CS 1.1 release #12
Integrate the following CLs from Perforce:

625267: Add new item in ini file to configure ApKeepAliveTimeout.
Currently We cannot configure the Ap Keep Alive Timeout. It is fixed
as 20 sec.  So new item is added in ini file to configure the Ap Keep
Alive Timeout.  Default value is 20 sec if this is not configured in
ini file. (CR 340193)

625405: Exit BMPS before creating the Second Session in case Of Concurrency.
Currently we are exiting BMPS only after creating the Second
Session. We are adding the Sta while we are in BMPS state. Because of
this System is going into wierd State. Now we are exiting BMPS before
creating the new session. (CR 347374)

625935: Link speed should reflect HT40 and SGI
It was reported that the link speed displayed in the Wifi GUI did not
reflect HT40 or SGI setting of the connected AP.  Analysis of the code
showed that SGI was being considered, but HT40 was not.  Update the
code to extract the HT40 setting from the statistics returned from
WCNSS.  Note that there is a companion change to WCNSS to properly
report the HT40 setting.  In addition update the logic used to
determine and cache the max speed since the max speed can change over
time. (CR 354157)

626069: Update OBSS config to Riva
OBSS feature enable/disable config is passed to Riva.  This field was
already there in HAL message, however WDI was not populating it.
Depending on this feature, Riva will decide whether to filter all the
beacons in SoftAP mode.  By default Riva is reading this config as
enabled, and because of this host is recieving beacons in SoftAP
mode. (CR 331169)

626074: PMC should allow IMPS only when all sessions are disconnected.
Fix Connect/disconnect crash during STA-SAP concurrency. (CR 352829)

626112: Fixes to address concurrency scenarios for Bt AMP
To address the following failing scenarios:
1. File transfer does not fall back to BR/EDR and terminates the
   transfer with Wi-Fi functionality
2. Enabling Wi-Fi hotspot on the remote device while AMP transfer is
   in progress, crashes the sender

The root cause for #1 was the deregister with HCI was happening even
before AMP disconnect event is sent to HCI. The fix is the make sure
disconnect is sent before deregister.

The root cause for #2 was we had pkts in WLAN stack on Rx while client
(HCI) has already deregistered, causing a NULL pointer access. The fix
is to add sanity check on rx path to take care of this scenario.  (CRs
355459, 355467)

626211: WLAN SETBAND failed in connected status
Fix error handling for 'wait_for_completion_interruptible_timeout' API
usage.  (CR 354562)

Change-Id: I5a6145a08827d800f5c8fdcdf5757f2b0b1e21cb