wlan: Fix for SAP turn on issue with dynamic ht_cap info.
With this change we will be dealing only with parameter present in the
ini file.  Its been found that SAP will not turn on if HT capability
info flags what we register and what are present in hostapd.conf are
not in sync. This CL is modification on CL#627756

Change-Id: I3f580e79b2a8a53b27f9c6285376caf8e52eb06a
CRs-Fixed: 363024