wlan: Fix RX skb corruption issue
On rare occasions skb corruption has been observed on the RX path.
This manifest itself as a kernel panic when the skb is freed after
having been processed by the kernel.  Analysis of the skb indicated
the skb_shared_info at the end of the buffer had been overwritten.
Further analysis showed the skb_shared_info was being overwritten when
frames were being DMA'ed into the buffer.  It was determined that
although the size of the RX buffers were large enough to handle normal
RX activity, in some error cases, particularly when an A-MSDU with
errors is processed, more than the expected number of bytes will be
transferred into the RX buffer.  Therefore we increase the packet
buffer size to accomodate the maximum sized frame that will be
transferred to the host.

Change-Id: Ic1da48e1e3bb653052dcf7c6fb96d37659b763ac
CRs-Fixed: 361728