wlan: Fix for sending proberequest on channel 165
channel list in WNI_CFG_SCAN_CONTROL_LIST is used for deciding whether
the proberequest need to be sent on this channel or not. 165 channel
in not there in this list due to this the proberequest is not being
transmitted on this channel. Added 165 channel to this list to fix
this issue.

Change-Id: I46595c18a7716bdc89e0976f0841632ac94f7546
CRs-Fixed: 362382