wlan: Private Ioctl to retrieve non-DFS and non-indoor channels for p2p
P2p requires a channel list with non-DFS and non-indoor channels so
that a list of possible operating channels for GO can be provided to
user space. Available channel list ioctl operates only in SAP/GO
mode. But we need this channel list before GO formation. So, an ioctl
in STA mode is implemented to get the channel list before GO
formation.

Change-Id: Ide408b4691dc663648b88823126e1f9b96ae9550
CRs-Fixed: 356831