wlan: Avoid leaking Management Frame Registrations
The memory allocation is happening in the
limProcessSmeRegisterMgmtFrameReq() function, where each allocated
tpSirRegisterMgmtFrame is inserted in the
&pMac->lim.gLimMgmtFrameRegistratinQueue at the time of driver
load. When the driver is unloaded the frames allocated above are
de-allocated in the same function.  The queue -
&pMac->lim.gLimMgmtFrameRegistratinQueue is destroyed in the
limCleanup() function. As a fail safe mechanism, to prevent any memory
leak, we can check the queue for any nodes present and deallocate them
before the queue is destroyed.

Change-Id: I5581fc753032eb65b821372d669d36ef26c6fd9f
CRs-Fixed: 354485