wlan: Add support for new Link Speed reporting mechanism
Previously our driver supported two algorithms for reporting link
speed, either reporting the current actual TX rate, or reporting the
maximum negotiated rate.  A third algorithm has been requested which
will factor in RSSI, so that feature request is incorporated as a
third link speed measurement algorithm.

The existing gReportMaxLinkSpeed configuration item is extended to
accept the following values:
0 - report the current actual TX rate
1 - report the maximum negotiated rate
2 - report the maximum negotiated rate with RSSI scaling

When gReportMaxLinkSpeed=2 the current RSSI is compared against two
threshold values to determine which link speed to report, If the RSSI
is greater than or equal to the "high" threshold (which is
configurable via the gLinkSpeedRssiHigh configuration item), then the
maximum negotiated rate is reported (in other words it acts exactly
like gReportMaxLinkSpeed=1).  If the RSSI is less than the "low"
threshold (which is configurable via the gLinkSpeedRssiLow
configuration item), then the current actual TX rate is reported (in
other words it acts exactly like gReportMaxLinkSpeed=0).  If the RSSI
falls between these two thresholds, then the maximum negotiated rate
is calculated, but it is capped at 5.5 Mbps for .11b connections, and
at 24 Mbps for .11a/.11g connections.

Note that in order to support the new RSSI configuration items, the
configuration parser is modified to support negative numbers.

Change-Id: Ie7bd5461e68faa1f08eb922ae9d30ec62b8532e8
CRs-Fixed: 357475