wlan: CS Release 1.7.5
Integrate the following CLs from Perforce:

630273:WLAN load fail cleanup
WLAN drive load failure crash the system. Close scheduler before
calling vos_close to make sure no thread is scheduled. (CR 361004)

630299: Fix for sending proberequest on channel 165
channel list in WNI_CFG_SCAN_CONTROL_LIST is used for deciding whether
the proberequest need to be sent on this channel or not. 165 channel
in not there in this list due to this the proberequest is not being
transmitted on this channel. Added 165 channel to this list to fix
this issue. (CR 362382)

630304: Remove HARD TABS from function __limHandleSmeStartBssRequest

630347: Change the prints that convey information about WCNSS version.
WCNSS version prints are not meaningful. Corrected the prints to
convey WCNSS version more meaningfully. (CRs 354013)

630355 & 630480: Private Ioctl to retrieve non-DFS and non-indoor
channels for p2p
P2p requires a channel list with non-DFS and non-indoor channels so
that a list of possible operating channels for GO can be provided to
user space. Available channel list ioctl operates only in SAP/GO
mode. But we need this channel list before GO formation. So, an ioctl
in STA mode is implemented to get the channel list before GO
formation. (CR 356831)

630469: Avoid leaking Management Frame Registrations
The memory allocation is happening in the
limProcessSmeRegisterMgmtFrameReq() function, where each allocated
tpSirRegisterMgmtFrame is inserted in the
&pMac->lim.gLimMgmtFrameRegistratinQueue at the time of driver
load. When the driver is unloaded the frames allocated above are
de-allocated in the same function.  The queue -
&pMac->lim.gLimMgmtFrameRegistratinQueue is destroyed in the
limCleanup() function. As a fail safe mechanism, to prevent any memory
leak, we can check the queue for any nodes present and deallocate them
before the queue is destroyed. (CR 354485)

630570: Add support for new Link Speed reporting mechanism
Previously our driver supported two algorithms for reporting link
speed, either reporting the current actual TX rate, or reporting the
maximum negotiated rate.  A third algorithm has been requested which
will factor in RSSI, so that feature request is incorporated as a
third link speed measurement algorithm.
The existing gReportMaxLinkSpeed configuration item is extended to accept the following values:
0 - report the current actual TX rate
1 - report the maximum negotiated rate
2 - report the maximum negotiated rate with RSSI scaling
When gReportMaxLinkSpeed=2 the current RSSI is compared against two
threshold values to determine which link speed to report, If the RSSI
is greater than or equal to the "high" threshold (which is
configurable via the gLinkSpeedRssiHigh configuration item), then the
maximum negotiated rate is reported (in other words it acts exactly
like gReportMaxLinkSpeed=1).  If the RSSI is less than the "low"
threshold (which is configurable via the gLinkSpeedRssiLow
configuration item), then the current actual TX rate is reported (in
other words it acts exactly like gReportMaxLinkSpeed=0).  If the RSSI
falls between these two thresholds, then the maximum negotiated rate
is calculated, but it is capped at 5.5 Mbps for .11b connections, and
at 24 Mbps for .11a/.11g connections.
Note that in order to support the new RSSI configuration items, the
configuration parser is modified to support negative numbers. (CR
357475)

630602: Fix for Hidden SSID - Soft AP
Ensure that the hidden SSID flag is populated in the config BSS
request (CR 361231)

630657: VOS bug API
Configurable VOS_BUG API to panic the system. If the PANIC_ON_BUG is
set, calling this API would result in system panic (CR 364050)

Change-Id: I6ec0cf83267d4ca669365f4784485ac8b17b27f0