wlan: Implemented "gModulatedDTIM" feature
Added new Cfg parameter to enable ModulatedDTIM in riva upon
suspend. Since setPowerParamsRequest is sent as part of suspend and
setPowerParams.uDTIMPeriod is unused, i used this variable to
communicat riva about this parameter. As part of resume,
setPowerParams is send to riva with uDTIMPeriod=0 so that this feature
gets disabled in riva.

Change-Id: I7ece4a274b211720c05f0840b782690bde56e713
CRs-Fixed: 359253