wlan: Fix of  data abort crashes in RIVA.
As part of DEL STA we are observing DATA ABORT RIVA crash while
reading RXP registers. The current theory behind the crash is Tx data
is ongoing while we are trying to delete STA context from HW and this
is leading to DATA ABORT. This is current best available solution and
team is working on finding the best solution of the problem.

Change-Id: I6cb51893423cfbdfe4961cebdf5efa20a77da5b8
CRs-Fixed: 361801