Volans WLAN driver changes to work with kernel 3.4
Changes to compile volans WLAN driver source against Kernel 3.4.
Changes are mainly to accomadate the cfg80211_ops API changes.
The softap config80211 interfaces are changed in Kernel 3.4,
some of the interfaces are renamed; and also a new data structure
is introduced for softap interface. So the existing WLAN driver
APIs are modified to accomadate these changes.

Change-Id: Ie51f583ab2a52db039a2a04499e71c5ff929513b