wlan: Add support for multicast MAC address list
Provision to configure multicast MAC address list "gMCAddrListEnable"
flag and support for Multicast MAC address List.  Currently Multicast
MAC address List is not supported. Nowit is supported in Riva and Host
changes are verified. In addition to this, one ini flag
"gMCAddrListEnable" is introduced to enable this support.
  # to enable Multicast MAC address list
  # Flags to enable filter Mcast MAC address list separately
  # gMCAddrListEnable=1 to enable the support
  # gMCAddrListEnable=0 to disable the support
  gMCAddrListEnable=1

Default value 0 if not configured in ini file.

Change-Id: I11b2f638788229012a9524b6d482adf18dbbff60
CRs-Fixed: 371401