wlan: Avoid unwanted reset of country code after disassoc/assoc failure
After dis-association or association failure, reset the country code
only if 11d is disabled and current operating country code and default
country codes are different.  If we reset the country code, driver is
flushing the scan results and which is triggering full scan. Full scan
is taking 3-4 sec and this delay is causing connection issues in
roaming.

Change-Id: I4a087d218d21ee5597f6ddb46675e160a4c8952f