wlan: changes to improve the wifi direct connection
After sending any action frame the device has to
wait on that particular channel for at least 100ms to
receive the response for it. Changes done to extend the
remain on channel time in Tx and Rx path of action frame
in case if sufficient time is not available to receive the
response.Changes done to extend the remain on channel time in Tx
and Rx path of action frame in case if sufficient time is not
available to receive the response.

Change-Id: I09a9947d1f22dc4c29296dc7929bc7b046098654
CRs-Fixed: 374456