wlan: Conditional check to define OKC flag added
This change is originally part of Legacy Fast Roaming, LFR requires a
kernel patch to support WPA2-Enterprise-OKC. This kernel patch is
already included in open source kernel code, but the same changes are
not yet available in ics/ics_chocolate AU git which leads to compile
error. Hence this flag is added to turn off OKC on the non-patched
kernels. we need to turn on this flag on the kernels with patch
available

Change-Id: I9616ada18cc33b9869a447112511b91da5d9e4fd
CRs-Fixed: 362782