wlan: Enhance debugging of some cfg80211 interface APIs
Although most of the cfg80211 interface APIs have ENTER() and EXIT()
debug statements, some of them do not.  Add ENTER() to those that do not
have them, and add EXIT() to ones that do not have them and which have
an appreciable amount of processing such that ENTER() does not provide
sufficient debugging.

Change-Id: Ic929b590e34cddb4f4c48b0dec26254c853e91af