wlan: Fix race condition in cfg80211 scan api
The 'scan pending' flag is reset before invoking the kernel API. So
there is a possible race condition where in the next scan request could
be invoked before invoking the Kernel API; and that would overwrite
the scan request stored in the adapter. The fix is to reset the flag,
only after invoking Kernel API.

Fixed some lineendings / space issues in other files.

Change-Id: I9d20f2e456ca231b3f9aaaa28f5af0ee8ba5664d
CRs-Fixed: 383267