wlan: Fix for APPs at high power when wlan disconnects in suspend mode.
Observed that when WLAN disconnects with AP in suspend mode,
scanning getting initiated from the supplicant this result in WLAN is
rejecting the suspend request as the scanning is going on(i.e not in
IMPS/BMPS). Android is requesting for suspend for 8 times and giving up,
This result in APP's remaining at high power for long time. To fix this
issue we are aqcuiring the wakelock before the scan starts and releasing
it after the finish scan.

Change-Id: Ic87cb37bab7d9aa0c6d955bdd9d534bcd3b06756
CRs-Fixed: 382547