wlan: VOSS ASSERT in WLANSAP_pmcFullPwrReqCB
During the SAP on/off testing if the SAP off is initiated before the
full power mode callback is triggered, sme_stop() will call all the
pending PMC callback's with status FAILURE because of this above issue
is observed.
Solution: In the full power call back function WLANSAP_pmcFullPwrReqCB()
failure status is not handled so removing the vos_assert.

Change-Id: Ib5303cc2422c907168f973103f3ffdcf87b5530b
CRs-Fixed: 387996