wlan: Fix for driver crash if the nv file is old.
The reason wlan driver crashes if the WCNSS_qcom_wlan_nv.bin file is old
is because the driver reads the nv items from the file and typecast the
read data to the structure nvEFSTable_t, but the structure has modified
to accommodate new rates in the latest driver and now since the
structure mismatch the data gets corrupted and while copying the nv
items which was read from nv files to the global variable pnvEFSTable
causing the crash. If the nv items in the file is different than the
sizeof(nvEFSTable_t) then only copy the default nv items and return.

Change-Id: I052abf1661df2bd85a458f10a24e10ceb27a99e8
CR-Fixed: 400823