wlan: Out-of-bound array indexing while preparing scan channel list.
While preparing scan channel list, an incorrect operator precedence
operation is leading to out-of-bound array indexing. Due to this we
populate bogus channel numbers (perhaps values from the stack) into the
channel list and attempt to scan as well. This is leading to several
issues both on the host (spending several seconds on scans) and Riva
(BMPS suspend failures, improper HAL configuration, etc). This is fixed
by addressing the operator precedence operation.

Change-Id: I70e1113cb27577c0c174446ccf1eb0ec5c84c287
CR-Fixed: 405567